hw04-auth push 3
1. Removed middleware/passport.js to keep only auth.js file.
2. Corrected index.js to contacts.js
3. Corrected login and logout in user.js
4. Moved logic from server.js to app.js
