+++ middleware/auth.js
@@ -1,6 +1,5 @@
 const passport = require("passport");
-const JwtStrategy = require("passport-jwt").Strategy;
-const ExtractJwt = require("passport-jwt").ExtractJwt;
+const { Strategy: JwtStrategy, ExtractJwt } = require("passport-jwt");
 const User = require("../service/schemas/user");
 
 const jwtOptions = {
@@ -12,7 +11,7 @@ passport.use(
   new JwtStrategy(jwtOptions, async (jwtPayload, done) => {
     try {
       const user = await User.findById(jwtPayload.id);
-      if (!user) {
+      if (!user || !user.token) {
         return done(null, false);
       }
       return done(null, user);
