--- middleware/passport.js
@@ -1,26 +1,0 @@
-const passport = require("passport");
-const JwtStrategy = require("passport-jwt").Strategy;
-const ExtractJwt = require("passport-jwt").ExtractJwt;
-const User = require("../service/schemas/user");
-const jwtOptions = {
-  jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-  secretOrKey: process.env.JWT_SECRET,
-};
-
-const jwtStrategy = new JwtStrategy(jwtOptions, async (jwtPayload, done) => {
-  try {
-    const user = await User.findById(jwtPayload.id);
-
-    if (!user) {
-      return done(null, false);
-    }
-
-    return done(null, user);
-  } catch (error) {
-    return done(error, false);
-  }
-});
-
-passport.use(jwtStrategy);
-
-module.exports = passport;
